pyproject.toml
[project]
name = "elliotshoney"
version = "0.1.0"
description = "Simple LangGraph agent with Notion integration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=1.0.3",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-google-genai>=3.0.0",
    "langchain-openai>=0.2.0",
    "langchain>=0.3.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "langchain-mistralai>=1.0.1",
]
